refactor: separate http 1 and 2
